+++ src/components/Categories.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Categories() {
+  return (
+    <div></div>
+  )
+}
